tests for required fields on commands and some logging updates

+++ qt-ozwdaemon/mqttcommands/deleteAllReturnRoute.cpp
@@ -3,13 +3,21 @@
 MqttCommand_DeleteAllReturnRoute::MqttCommand_DeleteAllReturnRoute(QObject *parent) :
     MqttCommand(parent)
 {
-    this->m_requiredFields << "node";
+    this->m_requiredIntFields << "node";
 }
 MqttCommand* MqttCommand_DeleteAllReturnRoute::Create(QObject *parent) {
     return new MqttCommand_DeleteAllReturnRoute(parent);
 }
 
 bool MqttCommand_DeleteAllReturnRoute::processMessage(QJsonDocument msg) {
+    if (!this->checkNode(msg, "node")) {
+        QJsonObject js;
+        js["status"] = "failed";
+        js["Error"] = "Invalid Node Number";
+        emit sendCommandUpdate(GetCommand(), js);
+        return false;
+    }
+
     QTOZWManager *mgr = getOZWManager();
     if (mgr->deleteAllReturnRoute(msg["node"].toInt())) {
         QJsonObject js;
